+++ adapters/telegram/index.ts
@@ -1,0 +1,14 @@
+"use strict";
+const { TelegramAdapter } = require("./inbound");
+const { createTelegramClient, splitByNewline } = require("./client");
+const { createTelegramRateLimiter, sendMessageWithRetry } = require("./rate");
+const { resolveTelegramConfig } = require("./config");
+
+export = {
+  TelegramAdapter,
+  createTelegramClient,
+  createTelegramRateLimiter,
+  sendMessageWithRetry,
+  splitByNewline,
+  resolveTelegramConfig,
+};